APEXMALHAR-2202 Moved accumulations to org.apache.apex.malhar.lib.window.accumulation.
