+++ library/src/main/java/org/apache/apex/malhar/lib/window/accumulation/Average.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.apex.malhar.lib.window.impl.accumulation;
+package org.apache.apex.malhar.lib.window.accumulation;
 
 import org.apache.apex.malhar.lib.window.Accumulation;
 import org.apache.commons.lang3.tuple.MutablePair;
