+++ library/src/main/java/org/apache/apex/malhar/lib/window/accumulation/TopN.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.apex.malhar.lib.window.impl.accumulation;
+package org.apache.apex.malhar.lib.window.accumulation;
 
 import java.util.Collections;
 import java.util.Comparator;
