+++ library/src/main/java/org/apache/apex/malhar/lib/window/accumulation/TopNByKey.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.apex.malhar.lib.window.impl.accumulation;
+package org.apache.apex.malhar.lib.window.accumulation;
 
 import java.util.Comparator;
 import java.util.HashMap;
