+++ library/src/test/java/org/apache/apex/malhar/lib/window/accumulation/FoldFnTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.apex.malhar.lib.window.impl.accumulation;
+package org.apache.apex.malhar.lib.window.accumulation;
 
 import org.junit.Assert;
 import org.junit.Test;
