+++ library/src/test/java/org/apache/apex/malhar/lib/window/accumulation/GroupTest.java
@@ -16,10 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.apex.malhar.lib.window.impl.accumulation;
+package org.apache.apex.malhar.lib.window.accumulation;
 
 import java.util.List;
-
 import org.junit.Assert;
 import org.junit.Test;
