+++ library/src/test/java/org/apache/apex/malhar/lib/window/accumulation/MaxTest.java
@@ -16,27 +16,38 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.apex.malhar.lib.window.impl.accumulation;
+package org.apache.apex.malhar.lib.window.accumulation;
 
-import java.util.Set;
-
+import java.util.Comparator;
 import org.junit.Assert;
 import org.junit.Test;
 
 /**
- * Test for {@link RemoveDuplicates}.
+ * Test for Max accumulation
  */
-public class RemoveDuplicatesTest
+public class MaxTest
 {
   @Test
-  public void RemoveDuplicatesTest()
+  public void MaxTest()
   {
-    RemoveDuplicates<Integer> rd = new RemoveDuplicates<>();
+    Max<Integer> max = new Max<>();
     
-    Set<Integer> accu = rd.defaultAccumulatedValue();
-    Assert.assertEquals(0, accu.size());
-    Assert.assertEquals(1, rd.accumulate(accu, 10).size());
-    Assert.assertEquals(2, rd.accumulate(accu, 11).size());
-    Assert.assertEquals(2, rd.accumulate(accu, 11).size());
+    Assert.assertEquals((Integer)5, max.accumulate(5, 3));
+    Assert.assertEquals((Integer)6, max.accumulate(4, 6));
+    Assert.assertEquals((Integer)5, max.merge(5, 2));
+  
+    Comparator<Integer> com = new Comparator<Integer>()
+    {
+      @Override
+      public int compare(Integer o1, Integer o2)
+      {
+        return -(o1.compareTo(o2));
+      }
+    };
+    
+    max.setComparator(com);
+    Assert.assertEquals((Integer)3, max.accumulate(5, 3));
+    Assert.assertEquals((Integer)4, max.accumulate(4, 6));
+    Assert.assertEquals((Integer)2, max.merge(5, 2));
   }
 }
