--- library/src/test/java/org/apache/apex/malhar/lib/window/impl/accumulation/AverageTest.java
@@ -1,41 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.apex.malhar.lib.window.impl.accumulation;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.apache.commons.lang3.tuple.MutablePair;
-
-/**
- * Test for {@link Average}.
- */
-public class AverageTest
-{
-  @Test
-  public void AverageTest()
-  {
-    Average ave = new Average();
-    MutablePair<Double, Long> accu = ave.defaultAccumulatedValue();
-    
-    for (int i = 1; i <= 10; i++) {
-      accu = ave.accumulate(accu, (double)i);
-    }
-    Assert.assertTrue(5.5 == accu.getLeft());
-  }
-}
